+++ src/felupe/assembly/expression/_bilinear.py
@@ -34,44 +34,36 @@ class BilinearForm:
     Parameters
     ----------
     v : BasisField
-        An object with basis functions (gradients) of a field.
-    grad_v : bool, optional (default is False)
-        Flag to use the gradient of ``v``.
+        Basis (shape) functions and gradients of a field.
     u : BasisField
-        An object with basis function (gradients) of a field.
-    grad_u : bool, optional (default is False)
-        Flag to use the gradient of ``u``.
+        Basis (shape) functions and gradients of a field.
     dx : ndarray or None, optional (default is None)
         Array with (numerical) differential volumes.
 
     """
 
-    def __init__(self, v, u, grad_v=False, grad_u=False, dx=None):
+    def __init__(self, v, u, dx=None, **kwargs):
         self.v = v
-        self.grad_v = grad_v
         self.u = u
-        self.grad_u = grad_u
         self.dx = dx
 
         self._form = IntegralFormCartesian(
-            None, v.field, self.dx, u.field, grad_v, grad_u
+            fun=None, v=v.field, dV=self.dx, u=u.field, **kwargs
         )
 
-    def integrate(self, weakform, args=(), kwargs={}, parallel=False, sym=False):
+    def integrate(self, weakform, kwargs={}, parallel=False, sym=False):
         r"""Return evaluated (but not assembled) integrals.
 
         Parameters
         ----------
         weakform : callable
-            A callable function ``weakform(v, *args, **kwargs)``.
-        args : tuple, optional
-            Optional arguments for callable weakform
+            A callable function ``weakform(v, u, **kwargs)``.
         kwargs : dict, optional
             Optional named arguments for callable weakform
-        parallel : bool, optional (default is False)
-            Flag to activate parallel threading.
-        sym : bool, optional (default is False)
-            Flag to active symmetric integration/assembly.
+        parallel : bool, optional
+            Flag to activate parallel threading (default is False).
+        sym : bool, optional
+            Flag to active symmetric integration/assembly (default is False).
 
         Returns
         -------
@@ -79,33 +71,33 @@ class BilinearForm:
             Integrated (but not assembled) matrix values.
         """
 
-        if self.grad_v:
-            v = self.v.grad
-        else:
-            v = self.v.basis
-
-        if self.grad_u:
-            u = self.u.grad
-        else:
-            u = self.u.basis
-
-        values = np.zeros((len(v), v.shape[-4], len(u), u.shape[-4], *u.shape[-2:]))
+        values = np.zeros(
+            (
+                len(self.v.basis),
+                self.v.basis.shape[-4],
+                len(self.u.basis),
+                self.u.basis.shape[-4],
+                *self.u.basis.shape[-2:],
+            )
+        )
 
         if not parallel:
-            for a, vbasis in enumerate(v):
+            for a, vbasis in enumerate(self.v.basis):
                 for i, vb in enumerate(vbasis):
-                    for b, ubasis in enumerate(u):
+                    for b, ubasis in enumerate(self.u.basis):
                         for j, ub in enumerate(ubasis):
                             if sym:
+                                v = type(self.v.basis)(vb, self.v.basis.grad[a, i])
+                                u = type(self.u.basis)(ub, self.u.basis.grad[b, j])
                                 if len(vbasis) * a + i <= len(ubasis) * b + j:
                                     values[a, i, b, j] = values[b, j, a, i] = (
-                                        weakform(vb, ub, *args, **kwargs) * self.dx
+                                        weakform(v, u, **kwargs) * self.dx
                                     )
 
                             else:
-                                values[a, i, b, j] = (
-                                    weakform(vb, ub, *args, **kwargs) * self.dx
-                                )
+                                v = type(self.v.basis)(vb, self.v.basis.grad[a, i])
+                                u = type(self.u.basis)(ub, self.u.basis.grad[b, j])
+                                values[a, i, b, j] = weakform(v, u, **kwargs) * self.dx
 
         else:
             idx_a, idx_i, idx_b, idx_j = np.indices(values.shape[:4])
@@ -123,21 +115,19 @@ class BilinearForm:
                     idx_j[mask],
                 )
 
-            def contribution(values, a, i, b, j, sym, args, kwargs):
+            def contribution(values, a, i, b, j, sym, kwargs):
+                v = type(self.v.basis)(self.v.basis[a, i], self.v.basis.grad[a, i])
+                u = type(self.u.basis)(self.u.basis[b, j], self.u.basis.grad[b, j])
                 if sym:
                     values[a, i, b, j] = values[b, j, a, i] = (
-                        weakform(v[a, i], u[b, j], *args, **kwargs) * self.dx
+                        weakform(v, u, **kwargs) * self.dx
                     )
 
                 else:
-                    values[a, i, b, j] = (
-                        weakform(v[a, i], u[b, j], *args, **kwargs) * self.dx
-                    )
+                    values[a, i, b, j] = weakform(v, u, **kwargs) * self.dx
 
             threads = [
-                Thread(
-                    target=contribution, args=(values, a, i, b, j, sym, args, kwargs)
-                )
+                Thread(target=contribution, args=(values, a, i, b, j, sym, kwargs))
                 for a, i, b, j in aibj
             ]